requirements.txt
Django==4.2.14
djangorestframework==3.14.0
psycopg2-binary==2.9.5
django-environ==0.11.2
pillow==10.4.0
django-jazzmin==3.0.0
drf-spectacular==0.27.2
PyJWT==2.7.0
djangorestframework-simplejwt==5.2.2
